Do not detach child process if the interpreter is none
If the child process is spawned with a shell and detach is true then the shell
window stays open on windows (even if windowHide is true)

Not sure why the detach is there, could be we can remove it for
all interpreters. It's meant to allow the child to continue after the parent
exits so possibly it is to handle the case where pm2 crashes and we want the child
processes to stay up...  If that's a thing.